+++ core/internals/tests/builtins.test.js
@@ -1,4 +1,4 @@
-const { add, sub, mul, div, pow } = require("../../../core/internals/builtins");
+const { add, sub, mul, div, pow } = require("../builtins");
 
 test("Add lazily adds two numbers", () => {
   const sum = add.call(1, 2);
