+++ core/tests/compiler.test.js
@@ -1,4 +1,4 @@
-const compiler = require("../../core/compiler");
+const compiler = require("../compiler");
 test("Compile code", () => {
   return compiler.compile("[1 2 3] -> print\n").then(out => {
     expect(typeof out).toBe("string");
