+++ core/tests/lexer.test.js
@@ -1,4 +1,4 @@
-const lexer = require("../../core/lexer");
+const lexer = require("../lexer");
 const testfile = "'Hello world' -> print";
 test("Lexed tokens are stored in array", () => {
   return lexer(testfile).then(output => {
